requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
djangorestframework==3.10.3
httpie==1.0.3
idna==2.8
mysqlclient==1.4.4
psycopg2==2.8.4
Pygments==2.4.2
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
